+++ QUANTAXIS/QAData/__init__.py
@@ -23,9 +23,9 @@
 # SOFTWARE.
 
 from QUANTAXIS.QAData.QASeriesStruct import QA_DataStruct_Series
-from QUANTAXIS.QAData.data_fq import QA_data_stock_to_fq, QA_data_make_qfq, QA_data_make_hfq
+from QUANTAXIS.QAData.data_fq import QA_data_stock_to_fq
 from QUANTAXIS.QAData.data_marketvalue import QA_data_calc_marketvalue, QA_data_marketvalue
-from QUANTAXIS.QAData.data_resample import QA_data_tick_resample, QA_data_min_resample, QA_data_day_resample
+from QUANTAXIS.QAData.data_resample import QA_data_tick_resample_1min, QA_data_tick_resample, QA_data_min_resample, QA_data_day_resample, QA_data_ctptick_resample
 from QUANTAXIS.QAData.QADataStruct import (QA_DataStruct_Index_day,
                                            QA_DataStruct_Index_min,
                                            QA_DataStruct_Stock_day,