+++ QUANTAXIS/QAData/base_datastruct.py
@@ -22,6 +22,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+import json
 import datetime
 import os
 import statistics
@@ -63,7 +64,7 @@ class _quotation_base():
         if 'volume' not in DataFrame.columns and 'vol' in DataFrame.columns:
             DataFrame = DataFrame.assign(volume=DataFrame.vol)
         # 🛠todo 判断DataFame 对象字段的合法性，是否正确
-        self.data = DataFrame.sort_index()
+        self.data = DataFrame.drop_duplicates().sort_index()
         self.data.index = self.data.index.remove_unused_levels()
         # 🛠todo 该变量没有用到， 是不是 self.type = marketdata_type ??
 
@@ -367,6 +368,11 @@ class _quotation_base():
         except:
             return None
 
+    @property
+    @lru_cache()
+    def ndarray(self):
+        return self.to_numpy()
+
     '''
     ########################################################################################################
     计算统计相关的
@@ -591,6 +597,12 @@ class _quotation_base():
         except Exception as e:
             raise e
 
+    def reset_index(self):
+        return self.data.reset_index()
+
+    def rolling(self, N):
+        return self.groupby('code').rolling(N)
+
     def plot(self, code=None):
 
         def kline_formater(param):
@@ -733,25 +745,55 @@ class _quotation_base():
 
     def reindex(self, ind):
         """reindex
-        
+
         Arguments:
             ind {[type]} -- [description]
-        
+
         Raises:
             RuntimeError -- [description]
             RuntimeError -- [description]
-        
+
         Returns:
             [type] -- [description]
         """
 
-        if isinstance(ind,pd.MultiIndex):
+        if isinstance(ind, pd.MultiIndex):
             try:
                 return self.new(self.data.reindex(ind))
             except:
                 raise RuntimeError('QADATASTRUCT ERROR: CANNOT REINDEX')
         else:
-            raise RuntimeError('QADATASTRUCT ERROR: ONLY ACCEPT MULTI-INDEX FORMAT')
+            raise RuntimeError(
+                'QADATASTRUCT ERROR: ONLY ACCEPT MULTI-INDEX FORMAT')
+
+    def reindex_time(self, ind):
+        if isinstance(ind, pd.DatetimeIndex):
+            try:
+                return self.new(self.data.loc[(ind, slice(None)), :])
+            except:
+                raise RuntimeError('QADATASTRUCT ERROR: CANNOT REINDEX')
+
+        else:
+            raise RuntimeError(
+                'QADATASTRUCT ERROR: ONLY ACCEPT DATETIME-INDEX FORMAT')
+
+    def iterrows(self):
+        return self.data.iterrows()
+
+    def iteritems(self):
+        return self.data.iteritems()
+
+    def itertuples(self):
+        return self.data.itertuples()
+
+    def abs(self):
+        return self.new(self.data.abs())
+
+    def agg(self, func, axis=0, *args, **kwargs):
+        return self.new(self.data.agg(func, axis=0, *args, **kwargs))
+
+    def aggregate(self, func, axis=0, *args, **kwargs):
+        return self.new(self.data.aggregate(func, axis=0, *args, **kwargs))
 
     def tail(self, lens):
         """返回最后Lens个值的DataStruct
@@ -787,7 +829,7 @@ class _quotation_base():
         """
         转换DataStruct为list
         """
-        return np.asarray(self.data).tolist()
+        return self.data.reset_index().values.tolist()
 
     def to_pd(self):
         """
@@ -799,7 +841,7 @@ class _quotation_base():
         """
         转换DataStruct为numpy.ndarray
         """
-        return np.asarray(self.data)
+        return self.data.reset_index().values
 
     def to_json(self):
         """
@@ -807,11 +849,17 @@ class _quotation_base():
         """
         return QA_util_to_json_from_pandas(self.data.reset_index())
 
-    def to_csv(self,*args,**kwargs):
+    def to_string(self):
+        return json.dumps(self.to_json())
+
+    def to_bytes(self):
+        return bytes(self.to_string(), encoding='utf-8')
+
+    def to_csv(self, *args, **kwargs):
         """datastruct 存本地csv
         """
 
-        self.data.to_csv(*args,**kwargs)
+        self.data.to_csv(*args, **kwargs)
 
     def to_dict(self, orient='dict'):
         """
@@ -843,8 +891,60 @@ class _quotation_base():
     #     return pd.concat(list(map(lambda x: func(
     #         self.data.loc[(slice(None), x), :], *arg, **kwargs), self.code))).sort_index()
 
+    def apply(self, func, *arg, **kwargs):
+        """func(DataStruct)
+
+        Arguments:
+            func {[type]} -- [description]
+
+        Returns:
+            [type] -- [description]
+        """
+
+        return func(self, *arg, **kwargs)
+
     def add_func(self, func, *arg, **kwargs):
+        """QADATASTRUCT的指标/函数apply入口
+
+        Arguments:
+            func {[type]} -- [description]
+
+        Returns:
+            [type] -- [description]
+        """
+
         return self.groupby(level=1, sort=False).apply(func, *arg, **kwargs)
+
+    def get_data(self, columns, type='ndarray', with_index=False):
+        """获取不同格式的数据
+
+        Arguments:
+            columns {[type]} -- [description]
+
+        Keyword Arguments:
+            type {str} -- [description] (default: {'ndarray'})
+            with_index {bool} -- [description] (default: {False})
+
+        Returns:
+            [type] -- [description]
+        """
+
+        res = self.select_columns(columns)
+        if type == 'ndarray':
+            if with_index:
+                return res.reset_index().values
+            else:
+                return res.values
+        elif type == 'list':
+            if with_index:
+                return res.reset_index().values.tolist()
+            else:
+                return res.values.tolist()
+        elif type == 'dataframe':
+            if with_index:
+                return res.reset_index()
+            else:
+                return res
 
     def pivot(self, column_):
         """增加对于多列的支持"""