+++ QUANTAXIS/QAData/data_resample.py
@@ -26,6 +26,71 @@ from datetime import time
 
 import pandas as pd
 
+def QA_data_tick_resample_1min(tick, type_='1min', if_drop=True):
+
+    """
+    tick 采样为 分钟数据
+    1. 仅使用将 tick 采样为 1 分钟数据
+    2. 仅测试过，与通达信 1 分钟数据达成一致
+    3. 经测试，可以匹配 QA.QA_fetch_get_stock_transaction 得到的数据，其他类型数据未测试
+    demo:
+    df = QA.QA_fetch_get_stock_transaction(package='tdx', code='000001', 
+                                           start='2018-08-01 09:25:00',
+                                           end='2018-08-03 15:00:00')
+    df_min = QA_data_tick_resample_1min(df)
+    """
+    tick = tick.assign(amount=tick.price*tick.vol)
+    resx = pd.DataFrame()
+    _dates = set(tick.date)
+
+    for date in sorted(list(_dates)):
+        _data = tick.loc[tick.date == date]
+        # morning min bar
+        _data1 = _data[time(9, 25): time(11, 30)].resample(
+            type_, closed='left', base=30, loffset=type_).apply({'price': 'ohlc', 'vol': 'sum', 'code': 'last', 'amount': 'sum'})
+        _data1.columns = _data1.columns.droplevel(0)
+        # do fix on the first and last bar
+        # 某些股票某些日期没有集合竞价信息，譬如 002468 在 2017 年 6 月 5 日的数据
+        if len(_data.loc[time(9, 25): time(9, 25)]) > 0:
+            _data1.loc[time(9, 31): time(9, 31), 'open'] = _data1.loc[time(9, 26): time(9, 26), 'open'].values
+            _data1.loc[time(9, 31): time(9, 31), 'high'] = _data1.loc[time(9, 26): time(9, 31), 'high'].max()
+            _data1.loc[time(9, 31): time(9, 31), 'low'] = _data1.loc[time(9, 26): time(9, 31), 'low'].min()
+            _data1.loc[time(9, 31): time(9, 31), 'vol'] = _data1.loc[time(9, 26): time(9, 31), 'vol'].sum()
+            _data1.loc[time(9, 31): time(9, 31), 'amount'] = _data1.loc[time(9, 26): time(9, 31), 'amount'].sum()
+        # 通达信分笔数据有的有 11:30 数据，有的没有
+        if len(_data.loc[time(11, 30): time(11, 30)]) > 0:
+            _data1.loc[time(11, 30): time(11, 30), 'high'] = _data1.loc[time(11, 30): time(11, 31), 'high'].max()
+            _data1.loc[time(11, 30): time(11, 30), 'low'] = _data1.loc[time(11, 30): time(11, 31), 'low'].min()
+            _data1.loc[time(11, 30): time(11, 30), 'close'] = _data1.loc[time(11, 31): time(11, 31), 'close'].values
+            _data1.loc[time(11, 30): time(11, 30), 'vol'] = _data1.loc[time(11, 30): time(11, 31), 'vol'].sum()
+            _data1.loc[time(11, 30): time(11, 30), 'amount'] = _data1.loc[time(11, 30): time(11, 31), 'amount'].sum()
+        _data1 = _data1.loc[time(9, 31): time(11, 30)]
+
+        # afternoon min bar
+        _data2 = _data[time(13,0): time(15,0)].resample(
+                type_, closed='left', base=30, loffset=type_).apply({'price': 'ohlc', 'vol': 'sum', 'code': 'last', 'amount': 'sum'})
+
+        _data2.columns = _data2.columns.droplevel(0)
+        # 沪市股票在 2018-08-20 起，尾盘 3 分钟集合竞价
+        if (pd.Timestamp(date) < pd.Timestamp('2018-08-20')) and (tick.code.iloc[0][0] == '6'):
+            # 避免出现 tick 数据没有 1:00 的值
+            if len(_data.loc[time(13, 0): time(13, 0)]) > 0:
+                _data2.loc[time(15, 0): time(15, 0), 'high'] = _data2.loc[time(15, 0): time(15, 1), 'high'].max()
+                _data2.loc[time(15, 0): time(15, 0), 'low'] = _data2.loc[time(15, 0): time(15, 1), 'low'].min()
+                _data2.loc[time(15, 0): time(15, 0), 'close'] = _data2.loc[time(15, 1): time(15, 1), 'close'].values
+        else:
+            # 避免出现 tick 数据没有 15:00 的值
+            if len(_data.loc[time(13, 0): time(13, 0)]) > 0:
+                _data2.loc[time(15, 0): time(15, 0)] = _data2.loc[time(15, 1): time(15, 1)].values
+        _data2 = _data2.loc[time(13, 1): time(15, 0)]
+        resx = resx.append(_data1).append(_data2)
+    resx['vol'] = resx['vol'] * 100.0
+    resx['volume'] = resx['vol']
+    resx['type'] = '1min'
+    if if_drop:
+        resx = resx.dropna()
+    return resx.reset_index().drop_duplicates().set_index(['datetime', 'code'])
+
 
 def QA_data_tick_resample(tick, type_='1min'):
     """tick采样成任意级别分钟线
@@ -36,21 +101,61 @@ def QA_data_tick_resample(tick, type_='1min'):
     Returns:
         [type] -- [description]
     """
-    tick=tick.assign(amount=tick.price*tick.vol)
+    tick = tick.assign(amount=tick.price*tick.vol)
     resx = pd.DataFrame()
     _temp = set(tick.index.date)
-    
+
     for item in _temp:
         _data = tick.loc[str(item)]
         _data1 = _data[time(9, 31):time(11, 30)].resample(
-            type_, closed='right', base=30, loffset=type_).apply({'price': 'ohlc', 'vol': 'sum','code':'last','amount':'sum'})
+            type_, closed='right', base=30, loffset=type_).apply({'price': 'ohlc', 'vol': 'sum', 'code': 'last', 'amount': 'sum'})
 
         _data2 = _data[time(13, 1):time(15, 0)].resample(
-            type_, closed='right', loffset=type_).apply({'price': 'ohlc', 'vol': 'sum','code':'last','amount':'sum'})
+            type_, closed='right', loffset=type_).apply({'price': 'ohlc', 'vol': 'sum', 'code': 'last', 'amount': 'sum'})
 
         resx = resx.append(_data1).append(_data2)
-    resx.columns=resx.columns.droplevel(0)
-    return resx.reset_index().drop_duplicates().set_index(['datetime','code'])
+    resx.columns = resx.columns.droplevel(0)
+    return resx.reset_index().drop_duplicates().set_index(['datetime', 'code'])
+
+
+def QA_data_ctptick_resample(tick, type_='1min'):
+    """tick采样成任意级别分钟线
+
+    Arguments:
+        tick {[type]} -- transaction
+
+    Returns:
+        [type] -- [description]
+    """
+
+    resx = pd.DataFrame()
+    _temp = set(tick.TradingDay)
+
+    for item in _temp:
+
+        _data = tick.query('TradingDay=="{}"'.format(item))
+        try:
+            _data.loc[time(20, 0): time(21, 0), 'volume'] = 0
+        except:
+            pass
+
+        _data.volume = _data.volume.diff()
+        _data = _data.assign(amount=_data.LastPrice*_data.volume)
+        _data0 = _data[time(0, 0):time(2, 30)].resample(
+            type_, closed='right', base=30, loffset=type_).apply({'LastPrice': 'ohlc', 'volume': 'sum', 'code': 'last', 'amount': 'sum'})
+
+        _data1 = _data[time(9, 0):time(11, 30)].resample(
+            type_, closed='right', base=30, loffset=type_).apply({'LastPrice': 'ohlc', 'volume': 'sum', 'code': 'last', 'amount': 'sum'})
+
+        _data2 = _data[time(13, 1):time(15, 0)].resample(
+            type_, closed='right', base=30, loffset=type_).apply({'LastPrice': 'ohlc', 'volume': 'sum', 'code': 'last', 'amount': 'sum'})
+
+        _data3 = _data[time(21, 0):time(23, 59)].resample(
+            type_, closed='left', loffset=type_).apply({'LastPrice': 'ohlc', 'volume': 'sum', 'code': 'last', 'amount': 'sum'})
+
+        resx = resx.append(_data0).append(_data1).append(_data2).append(_data3)
+    resx.columns = resx.columns.droplevel(0)
+    return resx.reset_index().drop_duplicates().set_index(['datetime', 'code']).sort_index()
 
 
 def QA_data_min_resample(min_data,  type_='5min'):
@@ -100,7 +205,6 @@ def QA_data_min_resample(min_data,  type_='5min'):
     CONVERSION = {'code': 'first', 'open': 'first', 'high': 'max', 'low': 'min', 'close': 'last', 'vol': 'sum', 'amount': 'sum'} if 'vol' in min_data.columns else {
         'code': 'first', 'open': 'first', 'high': 'max', 'low': 'min', 'close': 'last', 'volume': 'sum', 'amount': 'sum'}
     resx = pd.DataFrame()
-    
 
     for item in set(min_data.index.date):
         min_data_p = min_data.loc[str(item)]
@@ -109,8 +213,8 @@ def QA_data_min_resample(min_data,  type_='5min'):
         f = min_data_p['{} 13:00:00'.format(item):].resample(
             type_, closed='right', loffset=type_).apply(CONVERSION)
         resx = resx.append(d).append(f)
-    
-    return resx.dropna().reset_index().set_index(['datetime','code'])
+
+    return resx.dropna().reset_index().set_index(['datetime', 'code'])
 
 
 def QA_data_day_resample(day_data,  type_='w'):
@@ -136,7 +240,7 @@ def QA_data_day_resample(day_data,  type_='w'):
     CONVERSION = {'code': 'first', 'open': 'first', 'high': 'max', 'low': 'min', 'close': 'last', 'vol': 'sum', 'amount': 'sum'} if 'vol' in day_data.columns else {
         'code': 'first', 'open': 'first', 'high': 'max', 'low': 'min', 'close': 'last', 'volume': 'sum', 'amount': 'sum'}
 
-    return day_data.resample(type_, closed='right').apply(CONVERSION).dropna().reset_index().set_index(['date','code'])
+    return day_data.resample(type_, closed='right').apply(CONVERSION).dropna().reset_index().set_index(['date', 'code'])
 
 
 if __name__ == '__main__':