+++ QUANTAXIS/QAEngine/QABaseEventLoop.py
@@ -1,8 +1,0 @@
-# 
-
-import asyncio
-
-
-class QABaseEventLoop(asyncio.AbstractEventLoop):
-    def __init__(self):
-        pass